requirement.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-environ==0.11.2
celery==5.3.4
mysqlclient==2.2.0


Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.7
django-environ>=0.11.2
mysqlclient>=2.2.0
celery>=5.3.0
